[GR-37638] Ensure LinkAtBuildTimeSupport is available early on.
PullRequest: graal/11431
